requirements.txt
openai
python-dotenv
langchain
langchain_openai